load data from dashboard_state.json if present
