+++ src/hypofuzz/frontend/src/pages/NotFound.tsx
@@ -1,0 +1,3 @@
+export function NotFoundPage() {
+  return <h1>Page not found!</h1>
+}
