+++ src/hypofuzz/frontend/src/vite-env.d.ts
@@ -1,1 +1,2 @@
 /// <reference types="vite-plugin-svgr/client" />
+/// <reference types="vite/types/importMeta.d.ts" />
