+++ tox.ini
@@ -14,12 +14,14 @@ commands =
     mypy src/hypofuzz/
 
 [testenv:docs]
+passenv=
+    HYPOFUZZ_DOCS_OUTPUT_DIR
 description = Builds docs with sphinx
 deps =
     --requirement deps/docs.txt
 commands =
     pip install --editable .
-    sphinx-build -W --keep-going docs-src docs/docs {posargs}
+    sphinx-build -W --keep-going docs-src {env:HYPOFUZZ_DOCS_OUTPUT_DIR:docs/docs} {posargs}
 
 [testenv:{test,pytest7}]
 description = Runs pytest with posargs - `tox -e test -- -v` == `pytest -v`
